refactor: centralize answer evaluation
