+++ static/practice.ts
@@ -5,6 +5,7 @@ import type {
   PracticeState,
   PracticeResult,
 } from '@/types/question';
+import { evaluateAnswer, getCorrectAnswerText } from '@/utils/answerUtils';
 
 // Timer functionality removed - practice mode runs without time constraints
 
@@ -361,30 +362,8 @@ class PracticeManager {
       return;
     }
 
-    const isCorrect = this.evaluateAnswer(question, userAnswer);
+    const isCorrect = evaluateAnswer(question, userAnswer);
     this.revealAnswer(question, isCorrect);
-  }
-
-  private evaluateAnswer(question: Question, userAnswer: string): boolean {
-    if (!question || !userAnswer) {return false;}
-
-    if (question.is_free) {
-      return userAnswer.toLowerCase().trim() === question.correct_answer.toLowerCase().trim();
-    }
-    
-    if (question.is_calculation) {
-      const userNum = parseFloat(userAnswer);
-      const correctNum = question.correct_answer_number || parseFloat(question.correct_answer);
-      if (isNaN(userNum) || isNaN(correctNum)) {return false;}
-      
-      const tolerance = Math.abs(correctNum * 0.05); // 5% tolerance
-      return Math.abs(userNum - correctNum) <= tolerance;
-    }
-    
-    // Multiple choice
-    const correctAnswerNumber = question.correct_answer_number;
-    if (correctAnswerNumber === undefined || correctAnswerNumber === null) {return false;}
-    return parseInt(userAnswer) === correctAnswerNumber;
   }
 
   private revealAnswer(question: Question, isCorrect: boolean): void {
@@ -398,15 +377,8 @@ class PracticeManager {
     }
 
     if (answerEl) {
-      let correctAnswerText = question.correct_answer;
-      
-      // For multiple choice questions, find the correct answer text
-      if (!correctAnswerText && question.correct_answer_number && question.answers) {
-        const correctAnswer = question.answers.find(a => a.answer_number === question.correct_answer_number);
-        correctAnswerText = correctAnswer ? correctAnswer.text : 'N/A';
-      }
-      
-      answerEl.innerHTML = `<strong>Correct Answer:</strong> ${correctAnswerText || 'N/A'}${question.answer_unit ? ' ' + question.answer_unit : ''}`;
+      const correctAnswerText = getCorrectAnswerText(question);
+      answerEl.innerHTML = `<strong>Correct Answer:</strong> ${correctAnswerText}`;
       answerEl.style.display = 'block';
     }
 
@@ -447,7 +419,7 @@ class PracticeManager {
       const question = this.state.questions[i];
       const userAnswer = this.state.answers[i];
       
-      if (userAnswer && this.evaluateAnswer(question, userAnswer)) {
+      if (userAnswer && evaluateAnswer(question, userAnswer)) {
         correctCount++;
       }
     }
@@ -483,20 +455,6 @@ class PracticeManager {
 
   private populateSummaryTable(): void {
     // This method is no longer needed as summary functionality is moved to summary.ts
-  }
-
-  private getCorrectAnswerText(question: any): string {
-    if (question.is_calculation) {
-      return question.correct_answer || 'Calculation required';
-    } else {
-      // For multiple choice, show the text of the correct answer
-      const correctAnswerNumber = question.correct_answer_number;
-      if (correctAnswerNumber && question.answers) {
-        const correctAnswer = question.answers.find((a: any) => a.answer_number === correctAnswerNumber);
-        return correctAnswer ? correctAnswer.text : question.correct_answer || 'Unknown';
-      }
-      return question.correct_answer || 'Unknown';
-    }
   }
 
   private reviewQuestion(): void {
